Add swap layouts to compact layout
Gives a bit more control over how new panes spawn and where. Starts with
normal 2-pane split between picker and editor. When you spawn a new
pane, it splits horizontally with the editor, but the editor takes 75%
of the height. Spawning yet another pane will split it vertically with
the existing pane below the editor, and adjust the editor to take 50% of
the height