Breaking: Remove ZIDE_EDITOR_CMD_MODE and ZIDE_EDITOR_CD_CMD
In order to be a bit more "batteries included", this release removes
those env vars and replaces them wih a `getEditorCommand` function which
has a mapping of commands to common editors. This will make it so those
editors "just work" since they're the 90% use case, with the downside of
it not being configurable outside of these editors.

If new editor/command combos are found, they'll have to be added to the
mapping to get support.