Breaking: Stop writing zide env to disk
It's just causing bugs. The limitation seems to be that zellij really
wants all tabs in a session to have the same environment, making
switching layouts in different tabs of the same session challenging and
basically unsupported. If you want to start zide with a new layout, the
best option is to create a new zellij session.