Replace `yazi` and `lf` multi-config switching with automatic column layouts (#9)
This replaces the (very hacky) system for mapping multi-column layouts
to specific configs and replaces it with a much more flexible system
that uses the width of the terminal pane to determine the layout.

In `yazi`, this is done via the [`auto-layout.yazi`
plugin](./yazi/plugins/auto-layout.yazi), which is included inside
zide’s [`yazi/plugins`](./yazi/plugins) directory, and is initialized by
the [`yazi/init.lua`](./yazi/init.lua) file. In `lf`, we take advantage
of the `on-redraw` command, which can be added to the [`lfrc`
file](./lf/lfrc).

This now removes the ability to map layouts to a config file, but I’m
pretty sure nobody used that. It also means that the columns will update
while using the layout, such as taking a pane full screen or opening a
different layout inside an existing zide session.